add namespace validator
